+++ packages/google-vertex/src/google-vertex-image-model.ts
@@ -1,9 +1,9 @@
-import { ImageModelV1, JSONValue } from '@ai-sdk/provider';
+import { ImageModelV1, ImageModelV1CallWarning } from '@ai-sdk/provider';
 import {
   Resolvable,
-  postJsonToApi,
   combineHeaders,
   createJsonResponseHandler,
+  postJsonToApi,
   resolve,
 } from '@ai-sdk/provider-utils';
 import { z } from 'zod';
@@ -25,6 +25,9 @@ interface GoogleVertexImageModelConfig {
 export class GoogleVertexImageModel implements ImageModelV1 {
   readonly specificationVersion = 'v1';
 
+  // https://cloud.google.com/vertex-ai/generative-ai/docs/model-reference/imagen-api#parameter_list
+  readonly maxImagesPerCall = 4;
+
   get provider(): string {
     return this.config.provider;
   }
@@ -38,24 +41,31 @@ export class GoogleVertexImageModel implements ImageModelV1 {
     prompt,
     n,
     size,
+    aspectRatio,
+    seed,
     providerOptions,
     headers,
     abortSignal,
   }: Parameters<ImageModelV1['doGenerate']>[0]): Promise<
     Awaited<ReturnType<ImageModelV1['doGenerate']>>
   > {
-    if (size) {
-      throw new Error(
-        'Google Vertex does not support the `size` option. Use ' +
-          '`providerOptions.vertex.aspectRatio` instead. See ' +
-          'https://cloud.google.com/vertex-ai/generative-ai/docs/image/generate-images#aspect-ratio',
-      );
+    const warnings: Array<ImageModelV1CallWarning> = [];
+
+    if (size != null) {
+      warnings.push({
+        type: 'unsupported-setting',
+        setting: 'size',
+        details:
+          'This model does not support the `size` option. Use `aspectRatio` instead.',
+      });
     }
 
     const body = {
       instances: [{ prompt }],
       parameters: {
         sampleCount: n,
+        ...(aspectRatio != null ? { aspectRatio } : {}),
+        ...(seed != null ? { seed } : {}),
         ...(providerOptions.vertex ?? {}),
       },
     };
@@ -68,7 +78,7 @@ export class GoogleVertexImageModel implements ImageModelV1 {
       successfulResponseHandler: createJsonResponseHandler(
         vertexImageResponseSchema,
       ),
-      abortSignal: abortSignal,
+      abortSignal,
       fetch: this.config.fetch,
     });
 
@@ -76,6 +86,7 @@ export class GoogleVertexImageModel implements ImageModelV1 {
       images: response.predictions.map(
         (p: { bytesBase64Encoded: string }) => p.bytesBase64Encoded,
       ),
+      warnings,
     };
   }
 }