+++ packages/openai/src/openai-image-model.test.ts
@@ -34,15 +34,17 @@ describe('doGenerate', () => {
 
     await model.doGenerate({
       prompt,
-      n: 2,
+      n: 1,
       size: '1024x1024',
+      aspectRatio: undefined,
+      seed: undefined,
       providerOptions: { openai: { style: 'vivid' } },
     });
 
     expect(await server.getRequestBodyJson()).toStrictEqual({
       model: 'dall-e-3',
       prompt,
-      n: 2,
+      n: 1,
       size: '1024x1024',
       style: 'vivid',
       response_format: 'b64_json',
@@ -63,8 +65,10 @@ describe('doGenerate', () => {
 
     await provider.image('dall-e-3').doGenerate({
       prompt,
-      n: 2,
+      n: 1,
       size: '1024x1024',
+      aspectRatio: undefined,
+      seed: undefined,
       providerOptions: { openai: { style: 'vivid' } },
       headers: {
         'Custom-Request-Header': 'request-header-value',
@@ -88,11 +92,39 @@ describe('doGenerate', () => {
 
     const result = await model.doGenerate({
       prompt,
-      n: 2,
+      n: 1,
       size: undefined,
+      aspectRatio: undefined,
+      seed: undefined,
       providerOptions: {},
     });
 
     expect(result.images).toStrictEqual(['base64-image-1', 'base64-image-2']);
+  });
+
+  it('should return warnings for unsupported settings', async () => {
+    prepareJsonResponse();
+
+    const result = await model.doGenerate({
+      prompt,
+      n: 1,
+      size: '1024x1024',
+      aspectRatio: '1:1',
+      seed: 123,
+      providerOptions: {},
+    });
+
+    expect(result.warnings).toStrictEqual([
+      {
+        type: 'unsupported-setting',
+        setting: 'aspectRatio',
+        details:
+          'This model does not support aspect ratio. Use `size` instead.',
+      },
+      {
+        type: 'unsupported-setting',
+        setting: 'seed',
+      },
+    ]);
   });
 });