+++ packages/ai/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,15 @@
 
 ### Patch Changes
 
+- Updated dependencies [19a2ce7]
+- Updated dependencies [6337688]
+- Updated dependencies [8b422ea]
+  - ai@4.0.28
+
+## 0.0.1
+
+### Patch Changes
+
 - Updated dependencies [a56734f]
 - Updated dependencies [9589601]
 - Updated dependencies [e3cc23a]