+++ packages/xai/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ai-sdk/xai
 
+## 1.0.15
+
+### Patch Changes
+
+- Updated dependencies [19a2ce7]
+- Updated dependencies [19a2ce7]
+- Updated dependencies [43b37f7]
+- Updated dependencies [6337688]
+  - @ai-sdk/provider@1.0.4
+  - @ai-sdk/provider-utils@2.0.6
+  - @ai-sdk/openai-compatible@0.0.14
+
 ## 1.0.14
 
 ### Patch Changes